Fixed linker errors caused by GameBoard.hpp
