+++ Pazaak/Game.hpp
@@ -1,5 +1,5 @@
 #pragma once
-#include <SFML/Graphics.hpp>
+
 #include "MainMenu.hpp"
 #include "Match.hpp"
